+++ docs/src/changelog.md
@@ -26,7 +26,8 @@ _Nothing yet._
 ### Features
 
 - Add `Marshaler` trait for custom serialization and deserialization.
-- Add `TomlMarshaler` to save the stores as TOML files.
+- Add `JsonMarshaler` and `PrettyJsonMarshaler` to save the stores as JSON files.
+- Add `TomlMarshaler` and `PrettyTomlMarshaler` to save the stores as TOML files.
 - Add `StoreCollection::app_handle`.
 - Allow to destroy stores, also deleting their respective files.
 - Introduce unchecked variants for `get` methods (e.g. `Store::get_unchecked`).